rag-auditor/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
qdrant-client==1.12.1
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.3
